added Necromancer

+++ src/main/arena/combatants/enemies/HealBot.java
@@ -16,6 +16,6 @@ public class HealBot extends Combatant implements DefensiveAbility {
     public void ability(Combatant other) {
         mp -= mpCost;
         if (mp < 0) mp = 0;
-        other.heal(abilityStrength);
+        other.heal((int) abilityStrength);
     }
 }
